+++ CHANGELOG.md
@@ -7,9 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased] - ${maven.build.timestamp}
 ### Added
 - `quest compass` location now allows variables
-- variable support in `point` and `globalpoint` events and conditions
+- `point` and `globalpoint` events and conditions now support variables
+- `cancel` event now has `bypass` keyword to ignore conditions in Quest Canceler
 ### Changed
 - message.yml file was deleted and instead the lang folder now contains all translations
+- `cancel` event now respects by default canceler conditions
 ### Deprecated
 ### Removed
 - undocumented prefix feature in conversation